Marathon one (#51)
* refactor: add rel to the link in the sponsors component

* refactor: change setTimeout to setInterval in the useTimer for constant time updating

* feat: update getAllEvents controller

Moved the sorting by sortArrayByParams of data coming from the server from the frontend to
the backend, in the controller getAllEvents. There also added a check on
the format of data coming from the database, made it to be an array even
at one event.

* feat: add api.js file with rtk-queries

* refactor: add in the store eventsAPI reducer

* refactor: add rtk-query to the components, replace the home component

Moved the Home component to its own folder. Updated the path of the
Home component in App.
Add rtq-queries to the About, Advertising, Schedule, Timer,
Participants, Registration SingleEvent components and replaced it with
useSingleEvent hook with createAsyncThunk query.
Add to the Registration component useEffect hook for unmount data
participant from form

* del: useSingleEvent hook and mainSponsorsSlice

* refactor: add env variable for return_url stripe process

* refactor: add env variables to the stripe checkout, return components

* refactor: delete createAsyncThink from listEventsSlice and rename him

* refactor: del data.json and update gitignore